requirements.txt
streamlit==1.32.0
pypdf2==3.0.1
python-dotenv==1.0.1
pandas==2.2.1
numpy==1.26.4
sentence-transformers==2.5.1
chromadb==0.4.22
langchain==0.1.20
langchain-core==0.1.52
langchain-text-splitters==0.0.2
langchain-community==0.0.38
langchain-groq==0.1.9